endpoint for creating a spaces
