+++ controllers/onboardingController.js
@@ -122,68 +122,3 @@ export const createAccount = async (req, res) => {
   }
 };
 
-// Controller function to create a space with devices
-export const createSpaceWithDevices = async (req, res) => {
-  const { spaceName, Devices } = req.body;
-
-  // Basic validation to ensure spaceName and Devices are provided
-  if (!spaceName || !Array.isArray(Devices) || Devices.length === 0) {
-    return res
-      .status(400)
-      .json({ error: "spaceName and Devices are required" });
-  }
-
-  try {
-    const createSpaceQuery = fql`spaces.create(${{
-      spaceName,
-      devices: Devices,
-    }})`;
-
-    const result = await faunaClient.query(createSpaceQuery);
-    res.status(201).json({
-      message: "Space with devices created successfully",
-      data: result,
-    });
-  } catch (error) {
-    if (error instanceof ServiceError) {
-      console.error("Fauna error:", error);
-      return res.status(500).json({
-        error: "An error occurred while creating the space with devices",
-      });
-    }
-
-    console.error("Error creating space with devices:", error);
-    res.status(500).json({
-      error: "An error occurred while creating the space with devices",
-    });
-  }
-};
-
-// Controller function to fetch all spaces with devices
-export const getAllSpacesWithDevices = async (req, res) => {
-  try {
-    // FQL query to retrieve all spaces
-    const getAllSpacesQuery = fql`spaces.all()`;
-
-    const result = await faunaClient.query(getAllSpacesQuery);
-
-    res.status(200).json({
-      message: "Spaces with devices fetched successfully",
-      data: result.data, // assuming `result.data` contains the list of spaces
-    });
-  } catch (error) {
-    if (error instanceof ServiceError) {
-      console.error("Fauna error:", error);
-      return res.status(500).json({
-        error: "An error occurred while fetching the spaces with devices",
-      });
-    }
-
-    console.error("Error fetching spaces with devices:", error);
-    res.status(500).json({
-      error: "An error occurred while fetching the spaces with devices",
-    });
-  } finally {
-    faunaClient.close();
-  }
-};
