+++ router/onboarding.routes.js
@@ -3,8 +3,6 @@ import {
   createAccount,
   createOrganization,
   getAllAccounts,
-  createSpaceWithDevices,
-  getAllSpacesWithDevices,
   getOrganization,
 } from "../controllers/onboardingController.js";
 
@@ -14,7 +12,5 @@ router.route("/createAccount").post(createAccount);
 router.route("/create-organization").post(createOrganization);
 router.route("/accounts").get(getAllAccounts);
 router.route("/organization").get(getOrganization);
-router.route("/createSpace").post(createSpaceWithDevices);
-router.route("/spaces").get(getAllSpacesWithDevices);
 
 export default router;
